Stabilize `const_refs_to_static`
update tests

fix bitwidth-sensitive stderr output

use build-fail for asm tests

+++ tests/ui/consts/issue-17718-const-bad-values.rs
@@ -1,10 +1,15 @@
+//@ normalize-stderr-32bit: "\(size: \d+, align: \d+\)" -> "(size: $$PTR, align: $$PTR)"
+//@ normalize-stderr-64bit: "\(size: \d+, align: \d+\)" -> "(size: $$PTR, align: $$PTR)"
+//@ normalize-stderr-test: "([0-9a-f][0-9a-f] |╾─*A(LLOC)?[0-9]+(\+[a-z0-9]+)?(<imm>)?─*╼ )+ *│.*" -> "HEX_DUMP"
+
 #![allow(static_mut_refs)]
 
 const C1: &'static mut [usize] = &mut [];
 //~^ ERROR: mutable references are not allowed
 
-static mut S: usize = 3;
-const C2: &'static mut usize = unsafe { &mut S };
-//~^ ERROR: referencing statics in constants
+static mut S: i32 = 3;
+const C2: &'static mut i32 = unsafe { &mut S };
+//~^ ERROR: it is undefined behavior to use this value
+//~| reference to mutable memory
 
 fn main() {}
